Rename dashboard scripts file
